fix: number of analyzed indexes in F005 fixed
After comparing total indexes size in F005 and L001, it turned out, that
the sizes differ. Indexes size in F005 was less than it should be (the
size should be the same in both F005 and L001).
The cause of the problem was the wrong redundant condition when loading
indexes. It was found and removed.
Test fot check of indexeses size difference between L001 and F005
added.
Also, F005 report renamed from F005 Autovacuum: Index Bloat (Estimated)
to F005 Autovacuum: Btree Index Bloat (Estimated) and an additional
notice added to the top of F005 report.